paperboy/pyproject.toml
[tool.poetry]
name = "paperboy"
version = "0.1.0"
description = ""
authors = ["Sammy Hajhamid <sh@pepsi.pw>"]
readme = "README.md"
package-mode = false
packages = [{ include = "paperboy" }]

[tool.poetry.dependencies]
python = "^3.13"
python-telegram-bot = { extras = ["callback-data"], version = "^21.10" }
python-dotenv = "^1.0.1"
pycups = "^2.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
